fix: resolve gci/gofmt/unconvert linting issues in commit_timestamp feature
- commit_timestamp_test.go: fix gci import order, move testify/assert into
  the third-party group (standard → third-party → milvus-io)
- compaction_trigger_test.go: fix gofmt comment alignment
- import_checker.go: remove unnecessary uint64() conversion on typeutil.Timestamp
  (typeutil.Timestamp is already defined as = uint64)

+++ internal/datacoord/import_checker.go
@@ -471,7 +471,7 @@ func (c *importChecker) unsetSegmentImporting(originSegmentIDs, statsSegmentIDs 
 	for _, segmentID := range isImportingSegments {
 		err := c.meta.UpdateSegmentsInfo(c.ctx,
 			UpdateIsImporting(segmentID, false),
-			UpdateCommitTimestamp(segmentID, uint64(nowTs)),
+			UpdateCommitTimestamp(segmentID, nowTs),
 		)
 		if err != nil {
 			log.Warn("update import segment failed", zap.Error(err))
